+++ juriscraper/opinions/united_states/state/ncctapp.py
@@ -11,7 +11,4 @@ from juriscraper.opinions.united_states.state import nc
 
 class Site(nc.Site):
     court = "coa"
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.court_id = self.__module__
+    court_number = 2